third_party/WebKit/LayoutTests/inspector-protocol/worker/exception-from-worker-contains-stack-expected.txt

CONSOLE ERROR: line 3: Uncaught Error
Tests that console message from worker contains stack trace.
Worker created
Worker created
Message with stack trace received.
